+++ odoo/local-src/stock_delivery_note/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_delivery_note
+from . import test_delivery_note_recipient